refactor: cardPolicy.ts를 cardCompany.ts와 cardBrand.ts로 분리
